# api/v1/swagger.yaml
...
swagger: "2.0"
...
info:
  version: 1.0.0
  title: Refocus API
  # description: >-
  #   A short description of the application. GFM syntax can be used for rich text
  #   representation.
  # termsOfService: The Terms of Service for the API.
  # contact:
  #   name: The identifying name of the contact person/organization.
  #   url: http://focus.salesforce.com/contact
  #   email: focusapi@salesforce.com
  # license:
  #   name: The license name used for the API.
  #   url: http://focus.salesforce.com/license

# =============================================================================
host: localhost:10010
...
basePath: /v1
...
schemes:
  - https
  - http

# =============================================================================
securityDefinitions:
  jwt:
    type: apiKey
    name: Authorization
    in: header
# =============================================================================
consumes:
  - application/json

# =============================================================================
produces:
  - application/json

# =============================================================================
paths:
...
  /samples:
    x-swagger-router-controller: samples
    get:
      security:
        - jwt: []
      summary: Find a sample or samples
      tags: [ samples ]
      description: >-
        Find a sample or samples. You may query using field filters with
        asterisk (*) wildcards. You may also optionally specify sort, limit,
        offset, and a list of fields to include in the response. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      externalDocs:
        description: >-
          GET /samples
        url: http://focus.salesforce.com/docs/api/samples
      operationId: findSamples
      parameters:
        -
          $ref: "#/parameters/SamplesFieldsParam"
        - name: sort
          in: query
          description: >
            Specify the sort order using a field name, e.g. '...?sort=name'. Prepend the
            field name with a minus sign to specify descending order, e.g.
            '...?sort=-name'.
          required: false
          type: array
          collectionFormat: csv
          items:
            type: string
            enum:
              - id
              - messageBody
              - messageCode
              - status
              - value
              - createdAt
              - updatedAt
              - -id
              - -messageBody
              - -messageCode
              - -status
              - -value
              - -createdAt
              - -updatedAt
              # TODO aspect name
              # TODO subject absolutePath
        -
          $ref: "#/parameters/limitParam"
        -
          $ref: "#/parameters/offsetParam"
      responses:
        200:
          description: >-
            Success, returns a list of samples.
          schema:
            type: array
            items:
              $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        401:
          $ref: "#/responses/401"
        default:
          $ref: "#/responses/genericError"
    post:
      security:
        - jwt: []
      summary: Create a new sample
      tags: [ samples ]
      description: >-
        Create a new sample. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: postSample
      parameters:
        -
          name: queryBody
          description: Request body
          in: body
          schema:
            type: object
            description: >
              An observation of a particular aspect for a particular subject at a
              particular point in time.
            properties:
              messageBody:
                type: string
                description: >
                  Optional message about the sample value. Informational only--not used to
                  determine the status of the sample. It's up to the lens whether/how to
                  display this, will typically be displayed only on hover or click.
                maxLength: 4096
              messageCode:
                type: string
                maxLength: 5
                description: >
                  Optional 5-charater code about the sample value. Informational only--not
                  used to determine the status of the sample. It's up to the lens
                  whether/how to display this, will typically be displayed without having
                  to hover or click.
              value:
                type: string
                maxLength: 255
                description: >
                  Stored as a string but will be treated as whatever the aspect’s
                  valueType was specified as, i.e. [BOOLEAN|NUMERIC|PERCENT]. If timeout
                  occurs, value is set to null.
              relatedLinks:
                type: array
                items:
                  $ref: "#/definitions/RelatedLinkRequest"
                description: >
                  RelatedLinks associated with this model.
              aspectId:
                type: string
                description: >
                  TODO
              subjectId:
                type: string
                description: >
                  TODO
            required:
              - aspectId
              - subjectId
      responses:
        201:
          description: >-
            Created.
          schema:
            $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /samples/{key}:
    x-swagger-router-controller: samples
    delete:
      security:
        - jwt: []
      summary: Delete the specified sample
      tags: [ samples ]
      description: >-
        Delete the specified sample. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: deleteSample
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the sample to delete.
          required: true
          type: string
      responses:
        200:
          description: >-
            Success
          schema:
            $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    get:
      security:
        - jwt: []
      summary: Retrieve the specified sample
      tags: [ samples ]
      description: >-
        Retrieve the specified sample by its id or name. You may also optionally
        specify a list of fields to include in the response. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: getSample
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the sample to retrieve
          required: true
          type: string
        -
          $ref: "#/parameters/SamplesFieldsParam"
      responses:
        200:
          description: >-
            Success.
          schema:
            $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    patch:
      security:
        - jwt: []
      summary: Update the specified sample
      tags: [ samples ]
      description: >-
        Update the specified sample. If a field is not included in the query
        body, that field will not be updated. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: patchSample
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the sample to update.
          required: true
          type: string
        -
          name: queryBody
          in: body
          required: true
          schema:
            type: object
            description: >
              An observation of a particular aspect for a particular subject at a
              particular point in time.
            properties:
              messageBody:
                type: string
                description: >
                  Optional message about the sample value. Informational only--not used to
                  determine the status of the sample. It's up to the lens whether/how to
                  display this, will typically be displayed only on hover or click.
                maxLength: 4096
              messageCode:
                type: string
                maxLength: 5
                description: >
                  Optional 5-charater code about the sample value. Informational only--not
                  used to determine the status of the sample. It's up to the lens
                  whether/how to display this, will typically be displayed without having
                  to hover or click.
              value:
                type: string
                maxLength: 255
                description: >
                  Stored as a string but will be treated as whatever the aspect’s
                  valueType was specified as, i.e. [BOOLEAN|NUMERIC|PERCENT]. If timeout
                  occurs, value is set to null.
              relatedLinks:
                type: array
                items:
                  $ref: "#/definitions/RelatedLinkRequest"
                description: >
                  RelatedLinks associated with this model.
      responses:
        200:
          description: >-
            Success.
          schema:
            $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    put:
      security:
        - jwt: []
      summary: Update the specified sample
      tags: [ samples ]
      description: >-
        Update the specified sample. If a field is not included in the query
        body, that field will be set to null or its default value. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: putSample
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the sample to update.
          required: true
          type: string
        -
          name: queryBody
          in: body
          required: true
          schema:
            type: object
            description: >
              An observation of a particular aspect for a particular subject at a
              particular point in time.
            properties:
              messageBody:
                type: string
                description: >
                  Optional message about the sample value. Informational only--not used to
                  determine the status of the sample. It's up to the lens whether/how to
                  display this, will typically be displayed only on hover or click.
                maxLength: 4096
              messageCode:
                type: string
                maxLength: 5
                description: >
                  Optional 5-charater code about the sample value. Informational only--not
                  used to determine the status of the sample. It's up to the lens
                  whether/how to display this, will typically be displayed without having
                  to hover or click.
              value:
                type: string
                maxLength: 255
                description: >
                  Stored as a string but will be treated as whatever the aspect’s
                  valueType was specified as, i.e. [BOOLEAN|NUMERIC|PERCENT]. If timeout
                  occurs, value is set to null.
              relatedLinks:
                type: array
                items:
                  $ref: "#/definitions/RelatedLinkRequest"
                description: >
                  RelatedLinks associated with this model.
      responses:
        200:
          description: >-
            Success.
          schema:
            $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /samples/upsert:
    x-swagger-router-controller: samples
    post:
      security:
        - jwt: []
      summary: Create or update a sample
      tags: [ samples ]
      description: >-
        Create or update a sample by name. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: upsertSample
      parameters:
        -
          name: queryBody
          description: Request body
          in: body
          schema:
            type: object
            description: >
              An observation of a particular aspect for a particular subject at a
              particular point in time.
            properties:
              messageBody:
                type: string
                description: >
                  Optional message about the sample value. Informational only--not used to
                  determine the status of the sample. It's up to the lens whether/how to
                  display this, will typically be displayed only on hover or click.
                maxLength: 4096
              messageCode:
                type: string
                maxLength: 5
                description: >
                  Optional 5-charater code about the sample value. Informational only--not
                  used to determine the status of the sample. It's up to the lens
                  whether/how to display this, will typically be displayed without having
                  to hover or click.
              name:
                type: string
                description: >
                  TODO
              value:
                type: string
                maxLength: 255
                description: >
                  Stored as a string but will be treated as whatever the aspect’s
                  valueType was specified as, i.e. [BOOLEAN|NUMERIC|PERCENT]. If timeout
                  occurs, value is set to null.
              relatedLinks:
                type: array
                items:
                  $ref: "#/definitions/RelatedLinkRequest"
                description: >
                  RelatedLinks associated with this model.
            required:
              - name
      responses:
        200:
          description: >-
            Success, returns created or updated sample.
          schema:
            $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /samples/upsert/bulk:
    x-swagger-router-controller: samples
    post:
      security:
        - jwt: []
      summary: Create or update multiple samples
      tags: [ samples ]
      description: >-
        Create or update multiple samples. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: bulkUpsertSample
      produces:
        - text/plain; charset=utf-8
      parameters:
        -
          name: queryBody
          description: Request body
          in: body
          schema:
            type: array
            items:
              type: object
              description: >
                An observation of a particular aspect for a particular subject at a
                particular point in time.
              properties:
                messageBody:
                  type: string
                  description: >
                    Optional message about the sample value. Informational only--not used to
                    determine the status of the sample. It's up to the lens whether/how to
                    display this, will typically be displayed only on hover or click.
                  maxLength: 4096
                messageCode:
                  type: string
                  maxLength: 5
                  description: >
                    Optional 5-charater code about the sample value. Informational only--not
                    used to determine the status of the sample. It's up to the lens
                    whether/how to display this, will typically be displayed without having
                    to hover or click.
                name:
                  type: string
                  description: >
                    TODO
                value:
                  type: string
                  maxLength: 255
                  description: >
                    Stored as a string but will be treated as whatever the aspect’s
                    valueType was specified as, i.e. [BOOLEAN|NUMERIC|PERCENT]. If timeout
                    occurs, value is set to null.
              required:
              - name
      responses:
        200:
          description: >-
            Success, returns OK
          schema:
            type: string
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /samples/{key}/relatedLinks:
    x-swagger-router-controller: samples
    delete:
      security:
        - jwt: []
      summary: Delete all the related links of the specified sample
      tags: [ samples ]
      description: >-
        Delete all the related links of the specified sample. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: deleteSampleRelatedLinks
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the sample for which related links needs to be
            deleted.
          required: true
          type: string
      responses:
        200:
          description: >-
            Success, returns the updated sample.
          schema:
            $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /samples/{key}/relatedLinks/{relName}:
    x-swagger-router-controller: samples
    delete:
      security:
        - jwt: []
      summary: Delete the specified related link of the specified sample. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      tags: [ samples ]
      description: >-
        Delete the specified related link of the specified sample.
      operationId: deleteSampleRelatedLinks
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the sample for which related link needs to be
            deleted.
          required: true
          type: string
        -
          name: relName
          in: path
          description: >-
            The name of the related link to delete.
          required: true
          type: string
      responses:
        200:
          description: >-
            Success, returns the updated sample.
          schema:
            $ref: "#/definitions/SamplesResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
...
definitions:
  AuthenticationError:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors
  BadRequest:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors
  ErrorResponse:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors
  #  required:
  #    - message
  HATEOAS:
    type: object
    description: >
      Hypertext As The Engine Of Application State.
    properties:
      href:
        type: string
        maxLength: 2082
        readOnly: true
        description: >
          URL of the related HATEOAS link you can use for subsequent calls.
      rel:
        type: string
        maxLength: 255
        readOnly: true
        description: >
          Link relation that describes how this link relates to the previous call.
      method:
        type: string
        readOnly: true
        maxLength: 255
        description: >
          The HTTP method required for the related call.
    required:
      - href
      - rel
      - method
  RelatedLinkResponse:
    type: object
    description: >
      RelatedLink that is associated with a model.
    properties:
      name:
        type: string
        maxLength: 25
        readOnly: true
        description: >
          Name of the related link.
      url:
        type: string
        readOnly: true
        maxLength: 4096
        description: >
          The relatedLink url. For example, a related link associated with a
          sample can have a link to the source from which that sample
          information is obtained.
  RelatedLinkRequest:
    type: object
    description: >
      RelatedLink that is associated with a model.
    properties:
      name:
        type: string
        minLength: 1
        maxLength: 25
        pattern: ^[0-9A-Za-z_\\ -]{1,25}$
        description: >
          Name of the related link.
      url:
        type: string
        minLength: 1
        maxLength: 4096
        description: >
          The relatedLink url. For example, a related link associated with a
          sample can have a link to the source from which that sample
          information is obtained.
    required:
      - name
      - url
...
  SamplesResponse:
    type: object
    description: >
      An observation of a particular aspect for a particular subject at a
      particular point in time.
    properties:
      id:
        type: string
        readOnly: true
      messageBody:
        type: string
        readOnly: true
        description: >
          Optional message about the sample value. Informational only--not used to
          determine the status of the sample. It's up to the lens whether/how to
          display this, will typically be displayed only on hover or click.
        maxLength: 4096
      messageCode:
        type: string
        readOnly: true
        maxLength: 5
        description: >
          Optional 5-charater code about the sample value. Informational only--not
          used to determine the status of the sample. It's up to the lens
          whether/how to display this, will typically be displayed without having
          to hover or click.
      status:
        type: string
        enum:
          - Critical
          - Invalid
          - Timeout
          - Warning
          - Info
          - OK
        readOnly: true
        description: >
          Status.
      previousStatus:
        type: string
        enum:
          - Critical
          - Invalid
          - Timeout
          - Warning
          - Info
          - OK
        readOnly: true
        description: >
          Previous Status.
      value:
        type: string
        readOnly: true
        maxLength: 255
        description: >
          Stored as a string but will be treated as whatever the aspect’s
          valueType was specified as, i.e. [BOOLEAN|NUMERIC|PERCENT]. If timeout
          occurs, value is set to null.
      provider:
        readOnly: true
        $ref: "#/definitions/UsersResponse"
        description: >
          User who provided this sample.
      createdAt:
        readOnly: true
        type: string
        description: >
          TODO
      updatedAt:
        readOnly: true
        type: string
        description: >
          TODO
      statusChangedAt:
        readOnly: true
        type: string
        description: >
          Time at which the status changed from one state to another.
      aspect:
        readOnly: true
        $ref: "#/definitions/AspectsResponse"
        description: >
          The aspect of this sample.
      relatedLinks:
        readOnly: true
        type: array
        items:
          $ref: "#/definitions/RelatedLinkResponse"
        description: >
          RelatedLinks associated with this model.
      apiLinks:
        readOnly: true
        type: array
        items:
          $ref: "#/definitions/HATEOAS"
        description: >
          Hypertext As The Engine Of Application State.
...
  UsersResponse:
    type: object
    description: >
      A person or API client who interacts with the Focus system.
    properties:
      email:
        type: string
        readOnly: true
        maxLength: 254
        description: >
          The user's email address.
      id:
        type: string
        readOnly: true
      imageUrl:
        type: string
        readOnly: true
        maxLength: 2082
        description: >
          Optional thumbnail image.
      name:
        type: string
        readOnly: true
        maxLength: 256
        description: >
          First and last name.
      password:
        type: string
        readOnly: true
      createdBy:
        type: string
        readOnly: true
        description: >
          TODO
      createdAt:
        type: string
        readOnly: true
        description: >
          TODO
      updatedAt:
        type: string
        readOnly: true
        description: >
          TODO
      # TODO profile!
      apiLinks:
        readOnly: true
        type: array
        items:
          $ref: "#/definitions/HATEOAS"
        description: >
          Hypertext As The Engine Of Application State.
...
parameters:
  limitParam:
    name: limit
    in: query
    description: >
        Maximum number of records to return in the response; also useful for
        pagination in the UI.
    required: false
    type: integer
  offsetParam:
    name: offset
    in: query
    description: >
        Offset for paginating the response.
    required: false
    type: integer
...
  SamplesFieldsParam:
    name: fields
    in: query
    description: >
          Comma-delimited list of field names to include in the response.
    required: false
    type: array
    collectionFormat: csv
    items:
      type: string
      enum:
        - id
        - messageBody
        - messageCode
        - provider
        - status
        - tags
        - value
        - createdAt
        - updatedAt
        - aspect
        - subject
...
responses:
  400:
    description: Bad Request
    schema:
      $ref: '#/definitions/BadRequest'
  401:
    description: Authentication Error
    schema:
      $ref: '#/definitions/AuthenticationError'
  403:
    description: Forbidden
    schema:
      $ref: '#/definitions/ErrorResponse'
  404:
    description: Resource Not Found
    schema:
      $ref: '#/definitions/ResourceNotFoundError'
  genericError:
    description: Error
    schema:
      $ref: '#/definitions/ErrorResponse'

# =============================================================================